#I47O1M 重构模组代码

+++ northstar-strategy-common/src/main/java/tech/xuanwu/northstar/strategy/common/model/entity/ModulePositionEntity.java
@@ -22,6 +22,6 @@ public class ModulePositionEntity {
 	
 	private int volume;
 	
-	private int multiplier;
+	private double multiplier;
 	
 }
